requirements.txt
Flask==2.3.3
numpy==1.26.4
Werkzeug==2.3.7
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78